refactor(track-peek): extract audio preview component
